Refactor ImagePickerConfirmationController to Swift
