+++ Wire-iOS/Sources/UserInterface/Components/Controllers/ImagePickerConfirmationController+UIImagePickerControllerDelegate.swift
@@ -18,58 +18,3 @@
 
 import Foundation
 
-private extension UIImage {
-
-    /// Fix the pngData method ignores orientation issue
-    var flattened: UIImage {
-        if imageOrientation == .up { return self }
-
-        return UIGraphicsImageRenderer(size: size, format: imageRendererFormat).image { _ in draw(at: .zero) }
-    }
-}
-extension ImagePickerConfirmationController: UIImagePickerControllerDelegate {
-
-    public func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
-        presentingPickerController = picker
-
-        guard let imageFromInfo = info[.editedImage] as? UIImage ?? info[.originalImage] as? UIImage else {
-            picker.dismiss(animated: true)
-            return
-        }
-
-        let image = imageFromInfo.flattened
-
-        switch picker.sourceType {
-        case .photoLibrary,
-             .savedPhotosAlbum:
-
-            let confirmImageViewController = ConfirmAssetViewController()
-            confirmImageViewController.modalPresentationStyle = .fullScreen
-            confirmImageViewController.image = image
-            confirmImageViewController.previewTitle = previewTitle
-
-
-            confirmImageViewController.onCancel = {
-                picker.dismiss(animated: true)
-            }
-
-            confirmImageViewController.onConfirm = { [weak self] editedImage in
-                if let editedImage = editedImage {
-                    self?.imagePickedBlock(editedImage.pngData())
-                } else {
-                    self?.imagePickedBlock(image.pngData())
-                }
-            }
-
-            picker.present(confirmImageViewController, animated: true)
-            picker.setNeedsStatusBarAppearanceUpdate()
-
-        case .camera:
-            picker.dismiss(animated: true)
-            imagePickedBlock(image.pngData())
-        @unknown default:
-            picker.dismiss(animated: true)
-            imagePickedBlock(image.pngData())
-        }
-    }
-}
